Crate nft_candy_machine

Modules

accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

nft_candy_machine_v2

program

utils

Structs

AddConfigLines

Add multiple config lines to the candy machine.

CandyMachine

Candy machine state and config data.

CandyMachineData

Candy machine settings data.

ConfigLine

Creator

EndSettings

GatekeeperConfig

Configurations options for the gatekeeper.

HiddenSettings

Hidden Settings for large mints used with offline data.

InitializeCandyMachine

Create a new candy machine.

MintNFT

Mint a new NFT pseudo-randomly from the config array.

UpdateCandyMachine

Update the candy machine state.

WhitelistMintSettings

WithdrawFunds

Withdraw SOL from candy machine account.

Enums

EndSettingType

ErrorCode

WhitelistMintMode

Constants

CONFIG_ARRAY_START

CONFIG_LINE_SIZE

Individual config line for storing NFT data pre-mint.

Statics

ID

The static program ID

Functions

check_id

Confirms that a given pubkey is equivalent to the program ID

entry

The Anchor codegen exposes a programming model where a user defines a set of methods inside of a #[program] module in a way similar to writing RPC request handlers. The macro then generates a bunch of code wrapping these user defined methods into something that can be executed on Solana.

entrypoint⚠

Safety

get_config_count

get_config_line

get_good_index

id

Returns the program ID

Type Definitions

Result

Anchor generated Result to be used as the return type for the program.