Module nft_candy_machine::accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

Structs

AddConfigLines

InitializeCandyMachine

MintNFT

UpdateCandyMachine

WithdrawFunds